feat: Enhance appointment management
- Use different component for patient and doctor appointment view
- Use a single path but conditionally load component based on user type
