+++ client/src/pages/DoctorAppointments.jsx
@@ -1,0 +1,139 @@
+import React, { useEffect, useState } from "react";
+import Empty from "../components/Empty";
+import Footer from "../components/Footer";
+import Navbar from "../components/Navbar";
+import fetchData from "../helper/apiCall";
+import { setLoading } from "../redux/reducers/rootSlice";
+import Loading from "../components/Loading";
+import { useDispatch, useSelector } from "react-redux";
+import { jwtDecode } from "jwt-decode";
+import axios from "axios";
+import toast from "react-hot-toast";
+import "../styles/user.css";
+
+const DoctorAppointments = () => {
+  const [appointments, setAppointments] = useState([]);
+  const dispatch = useDispatch();
+  const { loading } = useSelector((state) => state.root);
+  const { userId } = jwtDecode(localStorage.getItem("token"));
+
+  const getAllAppoint = async (e) => {
+    try {
+      dispatch(setLoading(true));
+      const temp = await fetchData(
+        `/appointment/getallappointments?search=${userId}`
+      );
+      setAppointments(temp);
+      dispatch(setLoading(false));
+    } catch (error) {}
+  };
+
+  useEffect(() => {
+    getAllAppoint();
+  }, []);
+
+  const complete = async (ele) => {
+    try {
+      await toast.promise(
+        axios.put(
+          "/appointment/completed",
+          {
+            appointid: ele?._id,
+            doctorId: ele?.doctorId?._id,
+            doctorname: `${ele?.userId?.firstname} ${ele?.userId?.lastname}`,
+          },
+          {
+            headers: {
+              Authorization: `Bearer ${localStorage.getItem("token")}`,
+            },
+          }
+        ),
+        {
+          success: "Appointment booked successfully",
+          error: "Unable to book appointment",
+          loading: "Booking appointment...",
+        }
+      );
+
+      getAllAppoint();
+    } catch (error) {
+      return error;
+    }
+  };
+
+  return (
+    <>
+      {loading ? (
+        <Loading />
+      ) : (
+        <section className="container notif-section">
+          <h2 className="page-heading">Your Appointments</h2>
+
+          {appointments.length > 0 ? (
+            <div className="appointments">
+              <table>
+                <thead>
+                  <tr>
+                    <th>S.No</th>
+                    <th>Doctor</th>
+                    <th>Patient</th>
+                    <th>Appointment Date</th>
+                    <th>Appointment Time</th>
+                    <th>Booking Date</th>
+                    <th>Booking Time</th>
+                    <th>Status</th>
+                    {userId === appointments[0].doctorId?._id ? (
+                      <th>Action</th>
+                    ) : (
+                      <></>
+                    )}
+                  </tr>
+                </thead>
+                <tbody>
+                  {appointments?.map((ele, i) => {
+                    return (
+                      <tr key={ele?._id}>
+                        <td>{i + 1}</td>
+                        <td>
+                          {ele?.doctorId?.firstname +
+                            " " +
+                            ele?.doctorId?.lastname}
+                        </td>
+                        <td>
+                          {ele?.userId?.firstname + " " + ele?.userId?.lastname}
+                        </td>
+                        <td>{ele?.date}</td>
+                        <td>{ele?.time}</td>
+                        <td>{ele?.createdAt.split("T")[0]}</td>
+                        <td>{ele?.updatedAt.split("T")[1].split(".")[0]}</td>
+                        <td>{ele?.status}</td>
+                        {userId === ele?.doctorId?._id ? (
+                          <td>
+                            <button
+                              className={`btn user-btn accept-btn ${
+                                ele?.status === "Completed" ? "disable-btn" : ""
+                              }`}
+                              disabled={ele?.status === "Completed"}
+                              onClick={() => complete(ele)}
+                            >
+                              Complete
+                            </button>
+                          </td>
+                        ) : (
+                          <></>
+                        )}
+                      </tr>
+                    );
+                  })}
+                </tbody>
+              </table>
+            </div>
+          ) : (
+            <Empty />
+          )}
+        </section>
+      )}
+    </>
+  );
+};
+export default DoctorAppointments;
